refactor(app): restructure packets query
