+++ app/src/lib/components/tables/packets.svelte
@@ -8,36 +8,12 @@ import { derived } from "svelte/store"
 import CellOriginChannel from "$lib/components/table-cells/cell-origin-channel.svelte"
 
 import ExplorerTablePaginated from "$lib/components/explorer-table-paginated.svelte"
-import { createQuery } from "@tanstack/svelte-query"
-import request from "graphql-request"
-import { packetsQuery } from "$lib/graphql/queries/packets"
+import { packetsQuery } from "$lib/queries/packets"
 
 // export let chains: Array<Chain>
 // export let pageSize: number // must be even
 
-const packets = createQuery({
-  queryKey: ["packets"],
-  refetchInterval: 5_000,
-  queryFn: async () => request(URLS.GRAPHQL, packetsQuery, {}),
-  select: data =>
-    data.v0_packets.map(packet => ({
-      source: {
-        chain_id: packet.from_chain_id ?? "unknown",
-        connection_id: packet.from_connection_id ?? "unknown",
-        channel_id: packet.from_channel_id ?? "unknown",
-        port_id: packet.from_port_id ?? "unknown"
-      },
-      destination: {
-        chain_id: packet.to_chain_id ?? "unknown",
-        connection_id: packet.to_connection_id ?? "unknown",
-        channel_id: packet.to_channel_id ?? "unknown",
-        port_id: packet.to_port_id ?? "unknown"
-      },
-      timestamp: packet.source_time,
-      destination_time: packet.destination_time
-    }))
-})
-
+let packets = packetsQuery()
 let packetsDataStore = derived(packets, $packets => $packets.data ?? [])
 
 type PacketRow = UnwrapReadable<typeof packetsDataStore>[number]
