+++ app/src/lib/graphql/queries/packets.ts
@@ -1,27 +1,9 @@
 import { graphql } from "gql.tada"
+import { packetListDataFragment } from "$lib/graphql/fragments/packets"
 
-export const packetListDataFragment = graphql(/* GraphQL */ `
-  fragment PacketListData on v0_packets {
-    from_chain_id
-    from_connection_id
-    from_channel_id
-    from_port_id
-    source_block_hash
-    source_time
-    to_chain_id
-    to_connection_id
-    to_channel_id
-    to_port_id
-    destination_block_hash
-    destination_time
-    source_data
-    status
-  }
-`)
-
-export const packetsQuery = graphql(
+export const packetsLatestQueryDocument = graphql(
   /* GraphQL */ `
-    query PacketsQuery($limit: Int = 100) {
+    query PacketsLatestQuery($limit: Int = 100) {
       v0_packets(limit: $limit, order_by: { source_time: desc_nulls_last }) {
         ...PacketListData
       }
